Move EhCacheProviderWithXMLOrJavaBasedConfig(Test) to org.ehcache.xml.provider package (#2876)
