+++ xml/src/test/resources/configs/disk-persistent-cache-with-heap-of-1-entry.xml
